requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
google-generativeai
edge-tts
openai-whisper
librosa
soundfile
spacy
sentence-transformers
textstat
transformers
torch
torchvision
diffusers>=0.30.1
accelerate
imageio-ffmpeg
huggingface_hub
fal-client
modal
opencv-python
Pillow
ultralytics
scikit-learn
xgboost
pandas
numpy
matplotlib
joblib
requests
httpx
aiofiles
pydantic
